requirements.txt
# Python dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
kagglehub>=0.2.0

# API
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# UI
streamlit>=1.27.0
plotly>=5.17.0
requests>=2.31.0

# Load testing
locust>=2.15.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
